Moved SFUtilities to Aggregator
