+++ Aggregator/SFUtilities.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using static System.Fabric.FabricClient;
 
-namespace ClusterCollector
+namespace Aggregator
 {
     public class SFUtilities
     {
